refactor transcription form
