+++ enjoy/src/renderer/components/transcriptions/transcription-create-form.tsx
@@ -1,0 +1,3 @@
+export const TranscriptionCreateForm = () => {
+  return <></>;
+};
